Fixed a FCGI.pm bug where HTTP headers are sent with LF not CRLF.
Per CGI/1.1 LF based separators are fine, but servers like lighttpd
gets confused when headers are separated with LF but the body contains
CRLF and mis-recognizes the boundary.

Thanks to fcharlier for the report and typester to help me debugging
nasty Test::Builder STDOUT autoflush that causes .t not be able to
reproduce this bug.