Reworked reloading. Now moved the watcher logic into Plack::Server (adapter) that has a watcher process and then a child process that runs the web server: HTTP::Server::PSGI.
This adds a new API to the Server adapter:

  $server->run_with_reload($builder, watch => \@watched_dirs);

where $builder is a code reference to return a PSGI application. This
is to delay the code compilation in the child process etc. so the
reloading actually works.

Also changed (reverted) the restarter to monitor in parent and run the
server in a forked child rather than the current way which is
opposite.