Improve text representation of stack trace.
There were a few issues when the original exception was an Exception::Class
object:

* Ghe text representation was not showing the original error message.
  Devel::StackTrace::AsHTML nicely adds the original exception text to the
  beginning of the HTML, but the text version did not. Instead the trace just
  showed the Exception::Class object getting passed around.

* The stack trace was not the original stack trace, but a new one.

So change the exception handler to see if the exception object is blessed and
has a `trace` method. If so, use that for the trace. This assumes any
exception object with a `trace` method returns a Devel::StackTrace object,
which I think is likely.

Then, change the way the message is displayed to add the original error
message, as AsHTML does, and to use Devel::StackTrace's `as_string()` method
rather than our own function that does the same thing.