Add tests to illustrate writability of query_parameters(), body_parameters() and parameters()
While all the tests are written to pass, they illustrate an
inconsistency in the API which could stand to be clarified in the code
or docs.

The hashref returned by query_parameters() and body_parameters() and
parameters() are all read/write -- subsequent calls to the same method
return the modified hashref.

However, modifying the hashes returned by body_parameters() or
query_paremeters() does not modify the hashref returned by
parameters(), which claims to be a merger of the two.

It seems that either all the return values should be read-only,
( always returning the same values ), or if modifying them is supported
then the parameters() hash should be updated when either of the
body_parameters() or query_parameters() hashes are updated.