Document for Middleware authors that a new object is not allocated per request.
This bit me because I wrote something to make working with the request fast and
convenient by caching the env and request in the MW object like so.

haz env =>
    isa    => 'HashRef';

haz request =>
    isa     => 'Plack::Request',
    default => sub {
        return Plack::Request->new($_[0]->env);
    };

haz parameters =>
    isa    => 'Hash::MultiValue',
    default => sub {
        return $_[0]->request->parameters;
    };

sub call {
    my $self = shift;
    $self->env(shift);

    my $foo = $self->parameters->{foo};
    ...
}

doy on #plack tried to argue this was "obvious" because MW is just sugar to create a
code ref, but that requires thinking about MW as sugar to make a code ref while the MW
interface and docs all treat it like an object.  It also requires knowing that $mw->to_app(@_)
is an object method and not a class method.  to_app() is deliberately blurily defined in
Plack::Component, so a mortal user would not be expected to know this.

Either way, it will save some other poor slob some time.