builder {} should return the PSGI app, not URLMap object.
This *might* break some backward compatiblity if a user expects
builder {} to return URLMap object when `mount "/"` is used inside the
builder block. The way it worked before was kind of inconsistent -
builder {} returns URLMap object if there's no middleware applied in
the root, but returns a PSGI code ref if there's any.

Usually plackup and PSGI web server automatically applies some
middleware and turned this URLMap object into an actual PSGI code ref,
but ideally the builder {} API should be just fixed to always return
the code reference.

If your code breaks with this, you were trying to use the undocumented
and inconsistently implemented feature of Plack::Builder.