Added a test and fix to avoid use of undefined warnings. Fixes #231
This is kind of a ghetto fix - the response_cb filter is "supposed to"
receive undef in the end to indicate that it reaches the end, BUT
there's currently no guarantee for the callback filter is called again
(call until it returns undef).

Fundamentally, the way we reuse the same chunk filter to apply to
ARRAY, IO objects and writer is prone to errors and we should probably
be more explicit.