Short circuit Plack::Loader->load() if loading Plack::Handler throws exceptions.
This fixes a bug in perl 5.8 where require() puts entry in %INC even
if the compilation fails, which leads to Plack::Server:: fallback
unexpectedly works. For instance when FCGI.pm was not installed,
Plack::Server got loaded and eventually gave a runtime error.

This whole mess will be removed once ::Server fallback is really
deprecated and removed.