Fix parsing of QUERY_STRING for valueless keys.
For a given 'QUERY_STRING: foo&bar=baz' the 'foo' parameter is not
accessible in Plack::Request::query_parameters(). There are a number of
applications using this pattern as a short way of passing information to
the handling code. This patch changes Plack::Request behaviour so that
foo makes its way in the underlaying Hash::MultiValue object as a key
with the empty string value. Application developers can now test the
existence of that key.