Add middleware to fix bad IIS7 bad behaviour.
https://rt.cpan.org/Ticket/Display.html?id=78377

Only applies to IIS version 7, with keepalives enabled.

IIS will happily read a 3XX response, and do the redirect,
sending it's own body to the client. The body of your redirect is
never read, meaning that the next client connection gets this
prepended to the HTTP header, which results in it getting
something insane looking, and generating a 500 page.

Apply with something like:

$psgi_app = Plack::Middleware::Conditional->wrap(
    $psgi_app,
    builder => sub { Plack::Middleware::IIS7KeepAliveFix->wrap($_[0]) },
    condition => sub {
        my ($env) = @_;
        return unless $env->{SERVER_SOFTWARE} && $env->{SERVER_SOFTWARE} =~ m!IIS\/7\.[0-9]!; 1; }, );
    },
);