Plack::Request: add support for multiple cookies with the same name
Add support for multiple cookies with the same name to Plack::Request,
the current Plack::Request::cookies() implementation throws away all
but the first value for multiple cookies with the same name.

Add another all_cookies() function to the API that allows you to get
at all the cookies you'd get in HTTP_COOKIE.

Before this patch it wasn't possible to get to this data without
inspecting the raw data in $env, e.g. Apache2::Request allows you to
get to this data becuase the cookies it return are an APR::Table
object, so this makes it possible to support applications that need to
deal with multiple values for the same cookie name.

You get cookies like that if you've had some code that's set cookies
for .example.com and some for www.examplec.com, browsers will then
happily send you both cookies.