Plack::Handler::Apache1: add a notice about this module not having recent Apache2 patches
There have been numerous patches (e.g. the cleanup handler support)
for Apache2 that haven't been tested/ported to this module. Make a
note of this in the POD so that any users know what they're getting
themselves into.t

The wording in the documentation was improved by ranguard on #plack:

    09:53:28 < avar> and https://github.com/plack/Plack/pull/340 too, since nobody seems to have apache1 anymore.
    10:04:50 < ranguard> avar: that reads a bit odd to me... maybe: https://gist.github.com/3898457  ?