Plack::Handler::Apache2:  add a new plack.handler.apache.request.object key
Add a plack.handler.apache.request.object key which contains the
Apache request object, this allows the application to access the
original Apache request object. Originally I had used psgi.input for
this, but in some cases it's going to be overwritten to e.g.:

    bless( \*{'Plack::TempBuffer::PerlIO::$io'}, 'FileHandle' )

I think something like Plack::Request might be overwriting it, but in
any case we need a new key to access this since other things are going
to munge it.

This patch only adds a new key and doesn't change any existing ones,
it's best viewed with "git show -w".