Plack::Handler::Apache2: preserve a copy of %ENV for cleanup handlers
For reasons unknown we'll *sometimes* have an emptied out %ENV by the
time we run the cleanup phase, I can reproduce this by requesting any
URL in my application a few times.

With this patch applied to Plack::Handler::Apache2:

    @@ -90,9 +92,12 @@ sub call_app {
             die "Bad response $res";
         }

    +    print STDERR "Have done everything, have the env = " . Data::Dumper::Dumper(\%ENV);
    +
         if (@{ $env->{'psgix.cleanup.handlers'} }) {
             $r->push_handlers(
                 PerlCleanupHandler => sub {
    +                print STDERR "Running the cleanup handler, have the env = " . Data::Dumper::Dumper(\%ENV);
                     for my $cleanup_handler (@{ $env->{'psgix.cleanup.handlers'} }) {
                         $cleanup_handler->($env);
                     }

I'll always get the full %ENV when I dump it in "Have done
everything", but by the time I get to running the cleanup handlers
it's sometimes been completely emptied.

Work around this by preserving %ENV and restoring it if it gets lost,
this helps a lot with migrating existing Apache applications which
look at %ENV, and since we normally set %ENV with the "subprocess_env"
cal it's really confusing to only provide it some of the time.