Plack::Middleware::LogDispatch now stringifies objects.
This is something I poached from @autarch.  If you're passing (for
instance) a Throwable exception to Log::Dispatch it will *not* stringify
the exception for you.  This patch handles this case and also adds a
test to make sure that code refs are not stringified, since
Log::Dispatch does accept those as valid messages.