call $body->add() outside of loop for speed
Follow-up on Dancer/Dancer#1092

For the Dancer ticket, the root cause for the
performance problem ended being with HTTP::Body, but
it seems there is still a small gain to be had by
calling '->add()' outside of the loop (which makes
sense for huge files).

The question is, should I add a performance test
to characterize exactly how much we're gaining with
this patch, or is it good (or not) as-is?