Add ConditionalGet support to P::MW::Static
Plack::App::File adds Last-Modified headers to served static content.
However; if you use the pass through feature of
Plack::Middleware::Static, its not straight forward to wrap the
responses from this middleware with ConditionalGet without also wrapping
the inner application.

To resolve this, add a conditional_get attribute to Plack::Middleware::Static.
If that has a true value, the Plack::App::File instance is wrapped with
Plack::Middleware::ConditionalGET, allowing 304 Not Modified responses
to be returned for static content; but any requests that pass through to
the underlying application are not wrapped by ConditionalGET.